LG Issuance and Cash Margin stability fixes
